+++ UnitsNet/CustomCode/Quantities/Length.extra.cs
@@ -12,7 +12,7 @@ namespace UnitsNet
 {
     public partial struct Length
     {
-        private const double InchesInOneFoot = 12;
+        private const decimal InchesInOneFoot = 12;
 
         /// <summary>
         ///     Converts the length to a customary feet/inches combination.
@@ -32,9 +32,9 @@ namespace UnitsNet
         /// <summary>
         ///     Get length from combination of feet and inches.
         /// </summary>
-        public static Length FromFeetInches(double feet, double inches)
+        public static Length FromFeetInches(decimal feet, decimal inches)
         {
-            return FromInches(InchesInOneFoot*feet + inches);
+            return FromInches(InchesInOneFoot * feet + inches);
         }
 
         /// <summary>
@@ -103,7 +103,7 @@ namespace UnitsNet
             {
                 result = feet + inches;
 
-                if(negativeSignGroup.Length > 0)
+                if (negativeSignGroup.Length > 0)
                     result = -result;
 
                 return true;
@@ -116,55 +116,55 @@ namespace UnitsNet
         /// <summary>Get <see cref="Speed"/> from <see cref="Length"/> divided by <see cref="TimeSpan"/>.</summary>
         public static Speed operator /(Length length, TimeSpan timeSpan)
         {
-            return Speed.FromMetersPerSecond(length.Meters/timeSpan.TotalSeconds);
+            return Speed.FromMetersPerSecond(length.Meters / (decimal)timeSpan.TotalSeconds);
         }
 
         /// <summary>Get <see cref="Speed"/> from <see cref="Length"/> divided by <see cref="Duration"/>.</summary>
         public static Speed operator /(Length length, Duration duration)
         {
-            return Speed.FromMetersPerSecond(length.Meters/duration.Seconds);
+            return Speed.FromMetersPerSecond(length.Meters / duration.Seconds);
         }
 
         /// <summary>Get <see cref="Duration"/> from <see cref="Length"/> divided by <see cref="Speed"/>.</summary>
         public static Duration operator /(Length length, Speed speed)
         {
-            return Duration.FromSeconds(length.Meters/speed.MetersPerSecond);
+            return Duration.FromSeconds(length.Meters / speed.MetersPerSecond);
         }
 
         /// <summary>Get <see cref="Area"/> from <see cref="Length"/> times <see cref="Length"/>.</summary>
         public static Area operator *(Length length1, Length length2)
         {
-            return Area.FromSquareMeters(length1.Meters*length2.Meters);
+            return Area.FromSquareMeters(length1.Meters * length2.Meters);
         }
 
         /// <summary>Get <see cref="Volume"/> from <see cref="Area"/> times <see cref="Length"/>.</summary>
         public static Volume operator *(Area area, Length length)
         {
-            return Volume.FromCubicMeters(area.SquareMeters*length.Meters);
+            return Volume.FromCubicMeters(area.SquareMeters * length.Meters);
         }
 
         /// <summary>Get <see cref="Volume"/> from <see cref="Length"/> times <see cref="Area"/>.</summary>
         public static Volume operator *(Length length, Area area)
         {
-            return Volume.FromCubicMeters(area.SquareMeters*length.Meters);
+            return Volume.FromCubicMeters(area.SquareMeters * length.Meters);
         }
 
         /// <summary>Get <see cref="Torque"/> from <see cref="Force"/> times <see cref="Length"/>.</summary>
         public static Torque operator *(Force force, Length length)
         {
-            return Torque.FromNewtonMeters(force.Newtons*length.Meters);
+            return Torque.FromNewtonMeters(force.Newtons * length.Meters);
         }
 
         /// <summary>Get <see cref="Torque"/> from <see cref="Length"/> times <see cref="Force"/>.</summary>
         public static Torque operator *(Length length, Force force)
         {
-            return Torque.FromNewtonMeters(force.Newtons*length.Meters);
+            return Torque.FromNewtonMeters(force.Newtons * length.Meters);
         }
 
         /// <summary>Get <see cref="KinematicViscosity"/> from <see cref="Length"/> times <see cref="Speed"/>.</summary>
         public static KinematicViscosity operator *(Length length, Speed speed)
         {
-            return KinematicViscosity.FromSquareMetersPerSecond(length.Meters*speed.MetersPerSecond);
+            return KinematicViscosity.FromSquareMetersPerSecond(length.Meters * speed.MetersPerSecond);
         }
 
         /// <summary>Get <see cref="Pressure"/> from <see cref="Length"/> times <see cref="SpecificWeight"/>.</summary>
@@ -183,7 +183,7 @@ namespace UnitsNet
         /// <summary>
         ///     Construct from feet and inches.
         /// </summary>
-        public FeetInches(double feet, double inches)
+        public FeetInches(decimal feet, decimal inches)
         {
             Feet = feet;
             Inches = inches;
@@ -192,12 +192,12 @@ namespace UnitsNet
         /// <summary>
         ///     The feet value it was constructed with.
         /// </summary>
-        public double Feet { get; }
+        public decimal Feet { get; }
 
         /// <summary>
         ///     The inches value it was constructed with.
         /// </summary>
-        public double Inches { get; }
+        public decimal Inches { get; }
 
         /// <inheritdoc cref="ToString(IFormatProvider)"/>
         public override string ToString()
